Time Lock Example
An basic example of how to incorporate a time lock to a smart contract so funds or assets cannot be accessed until a specified amount of time has passed. A common use case for ICOs, crowdfunding, and token vesting. However, this example allows the user to set the duration, so it would not be appropriate for many of the previously mentioned examples without modification. Best used for staking in contracts that reward interest.